+++ buildSrc/src/main/kotlin/Versions.kt
@@ -15,12 +15,13 @@ object Versions {
     const val ANDROIDX_LIFECYCLE = "2.8.2"
     const val OKIO = "3.9.0"
     const val KSP = "2.0.0-1.0.21"
-    const val KOTLIN_POET = "1.17.0"
+    const val KOTLIN_POET = "1.18.1"
     const val ACCOMPANIST = "0.34.0"
+    const val COMPILE_TESTING_KSP = "1.6.0"
 
     object Android {
-        const val TARGET_SDK = 34
-        const val COMPILE_SDK = 34
+        const val TARGET_SDK = 35
+        const val COMPILE_SDK = 35
         const val MIN_SDK = 21
     }
     const val JVM_TOOLCHAIN = 17