+++ trikot-kword/kword-plugin/src/main/kotlin/com.mirego.kword/KWordExtension.kt
@@ -1,0 +1,16 @@
+package com.mirego.kword
+
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.api.provider.Property
+
+abstract class KWordExtension {
+    abstract val translationFile: RegularFileProperty
+
+    abstract val translationFiles: ConfigurableFileCollection
+
+    abstract val enumClassName: Property<String>
+
+    abstract val generatedDir: DirectoryProperty
+}